sqlflite done

+++ lib/services/favorite.dart
@@ -11,6 +11,7 @@ class FavoriteService {
 
   static Future<void> addToFavorites(
       String userId, ProductModel product) async {
+       
     await db
         .collection(CollectionConfig.userFavorite)
         .where("user_id", isEqualTo: userId)
@@ -20,6 +21,7 @@ class FavoriteService {
         .then(
       (value) {
         if (value.docs.isEmpty) {
+           print("adding product to favs");
           db
               .collection(CollectionConfig.userFavorite)
               .doc()
@@ -34,18 +36,29 @@ class FavoriteService {
   }
 
   Future<void> insertToLocalFavorites(String userId, String productId) async {
-    int response = await sqlDb.insertData(
-      'INSERT INTO favorites (userId, productId) VALUES(?, ?)',
-      [userId, productId],
-    );
-    print(response);
-    print("addddddddddddddddddddddddddddeddddddddddddddddddd");
+    List<Map> exists = await sqlDb.readData(
+        "Select * from favorites Where userId = '${userId}' AND productId = '${productId}' LIMIT 1");
+    if (exists.isEmpty) {
+      int response = await sqlDb.insertData(
+          'INSERT INTO favorites (userId, productId) VALUES("${userId}", "${productId}")');
+      print(response);
+      print("addddddddddddddddddddddddddddeddddddddddddddddddd");
+    } else {
+      print("Already Exists");
+    }
   }
 
   Future<void> deleteFromLocalFavorites(String userId, String productId) async {
-    int response = await sqlDb.deleteData(
-        "DELETE FROM favorites WHERE userId = $userId AND productId = $productId ");
-    print(response);
+    List<Map> exists = await sqlDb.readData(
+        "Select * from favorites Where userId = '${userId}' AND productId = '${productId}' LIMIT 1");
+    if (exists.isNotEmpty) {
+      print("fav Delelted");
+      int response = await sqlDb.deleteData(
+          "DELETE FROM favorites WHERE userId = '$userId' AND productId = '$productId' ");
+      print(response);
+    } else {
+      print("Row Doesnt Exist");
+    }
   }
 
   static Future<void> removeProductFromFavorite(
